Added configurable language for everything and more
This can read the default minecraft language files, but in order to
translate the messages, you need to add the message section from the
en_US.lang.
Also, spawn eggs will work now, the title of music discs will be
displayed in the info sections and "tipped arrow effect" is now named
"potion effect" and is displayed also if the product is a potion,
splashed potion or a lingering potion (and of course tipped arrows).
The message configuration is no longer found in the config.yml file, but
has gone to the .lang file, so you need to reconfigure every message!
As the Jsoup library is no longer needed, the filesize of the final jar
shrunk quite a bit.
Maybe there are some more changes, which I already forgot.